Restructuring
Restructured Search-Dotnet & Storage-PowerShell
