--- AzureSearch/Dotnet/metadata_storage_path/metadata_storage_path.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.1209
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "metadata_storage_path", "metadata_storage_path\metadata_storage_path.csproj", "{9E028793-FEE9-4BFB-9FE9-4571366BB470}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9E028793-FEE9-4BFB-9FE9-4571366BB470}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9E028793-FEE9-4BFB-9FE9-4571366BB470}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9E028793-FEE9-4BFB-9FE9-4571366BB470}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9E028793-FEE9-4BFB-9FE9-4571366BB470}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7578261F-CD0B-44CF-9D3C-A5A6D875CD8A}
-	EndGlobalSection
-EndGlobal
