--- AzureSearch/Dotnet/metadata_storage_path/metadata_storage_path/bin/Debug/netcoreapp2.1/metadata_storage_path.runtimeconfig.dev.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "additionalProbingPaths": [
-      "C:\\Users\\akpriya\\.dotnet\\store\\|arch|\\|tfm|",
-      "C:\\Users\\akpriya\\.nuget\\packages",
-      "C:\\Program Files\\dotnet\\sdk\\NuGetFallbackFolder"
-    ]
-  }
-}
