--- AzureSearch/Dotnet/metadata_storage_path/metadata_storage_path/bin/Debug/netcoreapp2.1/metadata_storage_path.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "netcoreapp2.1",
-    "framework": {
-      "name": "Microsoft.NETCore.App",
-      "version": "2.1.0"
-    }
-  }
-}
