+++ Storage/Powershell/ExportResourcesListToStorage.ps1
